Create removeloopfromlinkedlist.c
Output :
Cycle detected and removed.

-----
linked list : 1 -> 2 -> 3 -> 4 -> 2
The loop is detected at node 4, this is the mergepoint passed to removeCycle().
Another way to removeCycle() would be to use hash map that stores the addresses of the node in the linked list. If there already exists a node in the hash map that is the loop node, hence setting the NULL in the node's next would remove the loop from the list.